feat: global modal 생성 및 modal state 추가, create와 update 시 다른 모달이 뜨도록 수정
