+++ src/app/dev/page.tsx
@@ -2,7 +2,7 @@
 import Chart from "@/components/Chart";
 import { useState } from "react";
 import LoadingIndicator from "@/components/LoadingIndicator";
-import Modal from "@/components/Modal";
+import Modal from "@/components/Modal/AttendanceModal";
 import { useDispatch, useSelector } from "react-redux";
 import { openModal } from "@/redux/reducer/modalSlice";
 import { RootState } from "@/redux/store";
@@ -13,17 +13,31 @@ export default function Dev() {
   if (!DEV) return <></>;
 
   const dispatch = useDispatch();
-  const modalState = useSelector((state: RootState) => state.modal.isOpen);
 
-  const openModalHandler = () => {
-    dispatch(openModal());
+  const handleOpenEditModal = () => {
+    dispatch(
+      openModal({
+        modalType: "EditModal",
+        isOpen: true,
+      }),
+    );
+  };
+  const handleOpenBasicModal = () => {
+    dispatch(
+      openModal({
+        modalType: "AttendanceModal",
+        isOpen: true,
+      }),
+    );
   };
 
   return (
     <div className="mx-auto flex min-h-screen w-full max-w-screen-md flex-col items-center">
       {DEV ? <h2>use here wisely</h2> : null}
-      <button onClick={openModalHandler}>modal click</button>
-      {modalState === true ? <Modal /> : ""}
+      <div>
+        <button onClick={handleOpenEditModal}>에딧 모달 열기</button>
+        <button onClick={handleOpenBasicModal}>기본 모달 열기</button>
+      </div>
     </div>
   );
 }
