+++ src/components/Modal/AttendanceEditModal.tsx
@@ -1,0 +1,10 @@
+export default function AttendanceEditModal() {
+  return (
+    // edit attendance modal content
+    <>
+      <div className="fixed left-[50%] top-[50%] z-[999] translate-x-[-50%] translate-y-[-50%]">
+        Edit Modal
+      </div>
+    </>
+  );
+}
