+++ src/redux/reducer/modalSlice.ts
@@ -1,6 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 export const initialState = {
+  modalType: "",
   isOpen: false,
 };
 
@@ -8,7 +9,9 @@ const modalSlice = createSlice({
   name: "modal",
   initialState,
   reducers: {
-    openModal: (state) => {
+    openModal: (state, actions) => {
+      const { modalType } = actions.payload;
+      state.modalType = modalType;
       state.isOpen = true;
     },
     closeModal: (state) => {
